Optimize a local file

Optimize a local images using the reSmush.it API.

Note that the default parameters of the function outfile = file overwrites the local file. See Examples.

Usage

resmush_file(file, outfile = file, qlty = 92, verbose = FALSE)

Arguments

file

Path to a local file. reSmush can optimize the following image files:

png

jpg

gif

bmp

tif

outfile

Path where the optimized file would be store in your disk. By default, it would override the file specified in file.

qlty

Only affects jpg files. Integer between 0 and 100 indicating the optimization level. For optimal results use vales above 90.

verbose

Logical. If TRUE displays a summary of the results.

Value

Writes on disk the optimized file if the API call is successful. In any case, a (invisibly) data frame with a summary of the process is returned as well.

See also

Examples

# \donttest{
png_file <- system.file("extimg/example.png", package = "resmush")

# For example, writes to a temp file
tmp_png <- tempfile(fileext = ".png")

# Silently returns a data frame
png_res <- resmush_file(png_file, outfile = tmp_png)
png_res[, -c(1:2)]
#>   src_size dest_size compress_ratio notes
#> 1 239.9 Kb   70.7 Kb          29.5% OK ;)

# Use with jpg and parameters
jpg_file <- system.file("extimg/example.jpg", package = "resmush")

# For example, writes to a temp file
tmp_jpg <- tempfile(fileext = ".jpg")

# Silently returns a data frame
resmush_file(jpg_file, outfile = tmp_png, verbose = TRUE)
#> ✔ D:/a/_temp/Library/resmush/extimg/example.jpg optimized: 100.4 Kb => 83.2 Kb (82.8%)
#> ℹ Check output: C:\Users\RUNNER~1\AppData\Local\Temp\RtmpK8u2ma\file104c16b4684b.png
resmush_file(jpg_file, outfile = tmp_png, verbose = TRUE, qlty = 10)
#> ✔ D:/a/_temp/Library/resmush/extimg/example.jpg optimized: 100.4 Kb => 6.4 Kb (6.4%)
#> ℹ Check output: C:\Users\RUNNER~1\AppData\Local\Temp\RtmpK8u2ma\file104c16b4684b.png
# }