Optimize an online file

Optimize and download an online image using the reSmush.it API.

Usage

resmush_url(
  url,
  outfile = file.path(tempdir(), basename(url)),
  qlty = 92,
  verbose = FALSE,
  exif_preserve = FALSE
)

Arguments

url

url or a vector of urls pointing to hosted image files. reSmush can optimize the following image files:

png

jpg/jpeg

gif

bmp

tiff

webp

outfile

Path or paths where the optimized files would be store in your disk. By default, temporary files (see tempfile()) with the same basename() than the file provided in url would be created. It should be of the same length than url parameter.

qlty

Only affects jpg files. Integer between 0 and 100 indicating the optimization level. For optimal results use vales above 90.

verbose

Logical. If TRUE displays a summary of the results.

exif_preserve

Logical. Should the Exif information (if any) deleted? Default is to remove (i.e. exif_preserve = FALSE).

Value

Writes on disk the optimized file if the API call is successful. In any case, a (invisibly) data frame with a summary of the process is returned as well.

If any value of the vector outfile is duplicated, resmush_url() would rename the output with a suffix _1. _2, etc.

See also

Examples

# \donttest{

# Base url
base_url <- "https://raw.githubusercontent.com/dieghernan/resmush/main/inst/"

png_url <- paste0(base_url, "/extimg/example.png")
resmush_url(png_url)

# Several urls
jpg_url <- paste0(base_url, "/extimg/example.jpg")

summary <- resmush_url(c(png_url, jpg_url))

# Returns an (invisible) data frame with a summary of the process
summary
#>                                                                              src_img
#> 1 https://raw.githubusercontent.com/dieghernan/resmush/main/inst//extimg/example.png
#> 2 https://raw.githubusercontent.com/dieghernan/resmush/main/inst//extimg/example.jpg
#>                                                      dest_img src_size
#> 1 C:/Users/RUNNER~1/AppData/Local/Temp/Rtmpae8Wsq/example.png 239.9 Kb
#> 2 C:/Users/RUNNER~1/AppData/Local/Temp/Rtmpae8Wsq/example.jpg 100.4 Kb
#>   dest_size compress_ratio notes
#> 1   70.7 Kb          70.5% OK ;)
#> 2   83.2 Kb          17.2% OK ;)

# Use with jpg and parameters
resmush_url(jpg_url, verbose = TRUE)
#> ✔ Optimizing https://raw.githubusercontent.com/dieghernan/resmush/main/inst//extimg/example.jpg:
#> ℹ Effective compression ratio: 17.2%
#> ℹ Current size: 83.2 Kb (was 100.4 Kb)
#> ℹ Output: C:/Users/RUNNER~1/AppData/Local/Temp/Rtmpae8Wsq/example.jpg
resmush_url(jpg_url, verbose = TRUE, qlty = 10)
#> ✔ Optimizing https://raw.githubusercontent.com/dieghernan/resmush/main/inst//extimg/example.jpg:
#> ℹ Effective compression ratio: 93.6%
#> ℹ Current size: 6.4 Kb (was 100.4 Kb)
#> ℹ Output: C:/Users/RUNNER~1/AppData/Local/Temp/Rtmpae8Wsq/example.jpg
# }